#90 refactor: 타입 변경
