+++ src/components/Profile/ActivitySection/ActivityCard.tsx
@@ -10,20 +10,18 @@ import useUserInfoSuspenseQuery from '../hooks/useUserInfoSuspenseQuery';
 export const ActivityCard = ({
   title,
   icon,
+  loginedId,
   accessToken,
 }: ActivityCardProps) => {
-  const params = useParams();
-  console.log(params.userId);
+  const { userId } = useParams();
+
+  const currentProfileId = Number(userId) || Number(loginedId);
 
   const { data: currentUserInfo } = useUserInfoSuspenseQuery(
-    Number(params.userId),
+    currentProfileId,
     accessToken,
   );
 
-  // const currentUserInfo = useUserInfoStore((state: any) => state.userInfoData);
-  if (!currentUserInfo) {
-    return null;
-  }
   const { averageRating, reviewCount, mostFavoriteCategory } = currentUserInfo;
   return (
     <div className="flex flex-col flex-1 grow justify-center items-center gap-[20px] rounded-xl lg:max-w-[300px] lg:h-[128px] md:max-w-full md:h-[119px] mobile:w-[105px] mobile:h-[119px] bg-gray-25">
