+++ src/components/Profile/ProductSection/ProductSection.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { MenuDropdown } from '@/components/Profile/ProductSection/MenuDropdown';
 import { ProductList } from '@/components/Profile/ProductSection/ProductList';
 import { ProductMenuType } from '@/components/Profile/types/productType';
-import { PRODUCT_MENU } from '@/components/Profile/constants/PRODUCT_MENU';
+import { PRODUCT_MENU } from '@/components/Profile/constants/productMenu';
 import useProductsQuery from '@/components/Profile/hooks/useProductsQuery';
 
 interface ProductSectionProps {
@@ -27,11 +27,10 @@ export const ProductSection = ({ currentProfileId }: ProductSectionProps) => {
     <section className="flex flex-col gap-[30px]">
       <div className="!p-0 lg:hidden">
         <MenuDropdown
-          options={[
-            { value: '리뷰 남긴 상품', label: '리뷰 남긴 상품' },
-            { value: '등록한 상품', label: '등록한 상품' },
-            { value: '찜한 상품', label: '찜한 상품' },
-          ]}
+          options={PRODUCT_MENU.map((menu) => ({
+            value: menu,
+            label: menu,
+          }))}
           onSelect={(value: string) => {
             setActiveMenu(value as ProductMenuType);
           }}
