+++ src/components/Profile/types/userActivityType.ts
@@ -5,5 +5,6 @@ export type ActivityData = Omit<ActivityCardProps, 'accessToken'>;
 export interface ActivityCardProps {
   title: string;
   icon?: IconType;
+  loginedId?: string;
   accessToken: string;
 }
